+++ Code/CrowdControlMod.cs
@@ -937,6 +937,8 @@ public sealed class CrowdControlMod : Mod
         AddEffect(new ShuffleInventoryEffect());
         AddEffect(new SwitchLoadoutEffect());
         AddEffect(new NoItemPickupEffect(20));
+        AddEffect(new ItemMagnetEffect(true, 25));
+        AddEffect(new ItemMagnetEffect(false, 25));
         AddEffect(new ReforgeItemEffect());
         AddEffect(new MoneyBoostEffect(60));
         AddEffect(new GiveItemEffect(GiveItemEffect.GiveItem.Pickaxe));