+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "Give Yoyo" effect (disabled).
 - Added "Give Accessory" effect (disabled).
 - Added "Spawn Flying Piggy Bank" effect (disabled).
+- Added "Item Attractor" effect (disabled).
+- Added "Item Repeller" effect (disabled).
 
 ## [2.0.11] - 2023-12-13